+++ src/main/java/educanet/Player.java
@@ -6,11 +6,11 @@ public class Player {
 
     private String name;
     private String symbol;
-    private final int ID;
+    private final long ID;
 
-    private ArrayList<int[]> playHistory = new ArrayList<>(); //UNUSED -> maybe can speed up the checkWin method later
+    private ArrayList<long[]> playHistory = new ArrayList<>(); //UNUSED -> maybe can speed up the checkWin method later
 
-    public Player(String name, String symbol, int ID) {
+    public Player(String name, String symbol, long ID) {
         this.name = name;
         this.symbol = symbol;
         this.ID = ID;
@@ -33,16 +33,16 @@ public class Player {
         this.symbol = symbol;
     }
 
-    public int getID() {
+    public long getID() {
         return ID;
     }
 
-    public void savePlay(int[] pos) {
+    public void savePlay(long[] pos) {
         playHistory.add(pos);
     }
 
-    public void savePlay(int x, int y) {
-        savePlay(new int[]{x,y});
+    public void savePlay(long x, long y) {
+        savePlay(new long[]{x,y});
     }
